Update SparkLine.php
Fixes an issue where the SparkLine will only show SparkLineDay objects whose date does not differ from the current date by more than X days. Where X is the number of SparkLineDay objects passed into the SparkLine object (maxItemAmount).

That is because inside the resolveCoordinates method a collection of dates is created that is relative to the current date. Then those dates are used to find entries in the collection of SparkLineDay objects whose dates are not guaranteed to be relative to the current date.

This commit changed the way the SparkLineDay object of a day is retrieved by relying on the sort order of the $days collection.